requirements.txt
click==7.1.1
Flask==1.1.2
gitdb==4.0.4
GitPython==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
smmap==3.0.2
Werkzeug==1.0.1
